refactor(runner): replace fmtSize with shared notify.FormatBytes
