+++ internal/notify/unraid.go
@@ -10,6 +10,8 @@ import (
 	"os/exec"
 	"runtime"
 	"time"
+
+	"github.com/ruaan-deysel/vault/internal/format"
 )
 
 type Importance string
@@ -78,7 +80,7 @@ func Send(event, subject, description string, importance Importance) error {
 }
 
 func JobSuccess(jobName string, itemsDone int, sizeBytes int64) error {
-	desc := fmt.Sprintf("Backed up %d items (%s)", itemsDone, FormatBytes(sizeBytes))
+	desc := fmt.Sprintf("Backed up %d items (%s)", itemsDone, format.FormatBytes(sizeBytes))
 	return Send("Vault", fmt.Sprintf("Backup job '%s' completed", jobName), desc, ImportanceNormal)
 }
